% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidra-series-metadata.R
\name{.get_rate_series}
\alias{.get_rate_series}
\title{Get Rate Series Names}
\usage{
.get_rate_series()
}
\value{
Character vector of rate series names.
}
\description{
Returns series that are rates (typically derived from primary series).
}
\keyword{internal}
