% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidra-series-metadata.R
\name{.parse_sidra_period}
\alias{.parse_sidra_period}
\title{Map SIDRA Period String to YYYYMM Format}
\usage{
.parse_sidra_period(period_string)
}
\arguments{
\item{period_string}{Character. SIDRA period description.}
}
\value{
Integer. YYYYMM format of quarter end month.
}
\description{
Converts SIDRA period codes (e.g., "202301 janeiro - marco 2023") to
anomesfinaltrimmovel format (YYYYMM of quarter end month).
}
\keyword{internal}
