% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-sidra-population.R
\name{fetch-sidra-population}
\alias{fetch-sidra-population}
\title{Fetch Monthly Population from SIDRA}
\description{
Functions to download and transform IBGE's population estimates
from SIDRA API for use in monthly weight calibration.
}
\keyword{internal}
