% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-sidra-series.R
\name{fetch-sidra-series}
\alias{fetch-sidra-series}
\title{Fetch SIDRA Rolling Quarter Series}
\description{
Functions to download PNADC labor market indicators from IBGE's SIDRA API.
These series are published as rolling quarterly averages (trimestre movel).
}
\keyword{internal}
