% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-reference-month.R
\name{get_exception_quarters}
\alias{get_exception_quarters}
\title{Get Exception Quarters}
\usage{
get_exception_quarters()
}
\value{
Character vector of exception quarters in "YYYYtQ" format
}
\description{
Returns the list of quarters where IBGE used non-standard technical break timing.
In these quarters, the first reference week of a month only needs 3 days
(instead of the usual 4) within that month.
}
\details{
The exception quarters are documented in IBGE's methodology notes:
\itemize{
\item 2016t3: September 25 - October 1 technical break
\item 2016t4: December 25-31 technical break
\item 2017t2: June 25 - July 1 technical break
\item 2022t3: September 25 - October 1 technical break
\item 2023t2: June 25 - July 1 technical break
}
}
\examples{
get_exception_quarters()
# [1] "2016t3" "2016t4" "2017t2" "2022t3" "2023t2"

}
