% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidra-series-metadata.R
\name{get_sidra_categories}
\alias{get_sidra_categories}
\title{Get Series Categories}
\usage{
get_sidra_categories()
}
\value{
Character vector of valid category names.
}
\description{
Returns a list of valid series categories for filtering.
}
