% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidra-series-metadata.R
\name{list_sidra_series}
\alias{list_sidra_series}
\title{List Available SIDRA Series Names}
\usage{
list_sidra_series(category = NULL)
}
\arguments{
\item{category}{Optional character vector of categories to filter by.}
}
\value{
Character vector of series names.
}
\description{
Returns a character vector of all available series names that can be
fetched and mensalized.
}
\examples{
\dontrun{
# All series
list_sidra_series()

# Only population series
list_sidra_series(category = "population")
}

}
