% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensalize-sidra-series.R
\name{mensalize-sidra-series}
\alias{mensalize-sidra-series}
\title{Mensalize SIDRA Rolling Quarter Series}
\description{
Functions to convert IBGE's rolling quarterly (trimestre movel) series into
exact monthly estimates.
}
\keyword{internal}
