% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-reference-month.R
\name{print.pnadc_crosswalk}
\alias{print.pnadc_crosswalk}
\title{Print Method for pnadc_crosswalk}
\usage{
\method{print}{pnadc_crosswalk}(x, ...)
}
\arguments{
\item{x}{A pnadc_crosswalk object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Method for pnadc_crosswalk
}
